refactor(support)!: move npm-related helpers into appium package
`npm`/`NPM` in `@appium/support` was only ever consumed by the `appium`
package's CLI (extension install/update logic); no other package in the
monorepo or the wider appium-org used it. Move it to
`packages/appium/lib/utils/npm.ts`, merging in the near-identical
`resolveFrom` duplicate that already existed there and deduplicating the
`CACHE_DIR_RELATIVE_PATH`/lockfile constants against `lib/constants.ts`.

BREAKING CHANGE: `@appium/support` no longer exports `npm`, `ExecOpts`,
`InstallPackageOpts`, or `NpmInstallReceipt`.
