+++ packages/appium/lib/utils/index.ts
@@ -10,6 +10,8 @@ export {
 export {adler32} from './hash.js';
 export {isPackageChanged} from './is-package-changed.js';
 export type {IsPackageChangedOptions, IsPackageChangedResult} from './is-package-changed.js';
+export {npm, resolveFrom} from './npm.js';
+export type {ExecOpts, InstallPackageOpts, NpmExecResult, NpmInstallReceipt} from './npm.js';
 export {
   bindAll,
   camelCase,
@@ -29,4 +31,3 @@ export {
 } from './object.js';
 export {packageDidChange} from './package-changed.js';
 export {appiumPackageRoot, npmPackage} from './package-json.js';
-export {resolveFrom} from './resolve-from.js';
