+++ packages/appium/test/unit/utils/npm.spec.ts
@@ -2,19 +2,19 @@ import assert from 'node:assert/strict';
 import path from 'node:path';
 import {describe, it} from 'node:test';
 
-import {NPM, resolveFrom} from '../../lib/npm.js';
+import {NPM, resolveFrom} from '../../../lib/utils/npm.js';
 
-describe('npm', function () {
+describe('utils/npm', function () {
   describe('resolveFrom()', function () {
-    const supportRoot = path.join(import.meta.dirname, '..', '..');
+    const appiumRoot = path.join(import.meta.dirname, '..', '..', '..');
 
     it('should resolve a package path from a directory', async function () {
-      const resolved = await resolveFrom(supportRoot, 'semver/package.json');
+      const resolved = await resolveFrom(appiumRoot, 'semver/package.json');
       assert.match(resolved, /semver[/\\]package\.json$/);
     });
 
     it('should reject when the module cannot be resolved', async function () {
-      await assert.rejects(resolveFrom(supportRoot, 'nonexistent-appium-package-xyz/package.json'));
+      await assert.rejects(resolveFrom(appiumRoot, 'nonexistent-appium-package-xyz/package.json'));
     });
   });
